feat: code cleaned into modules
